requirements.txt
altair==5.5.0
asttokens==3.0.0
attrs==25.1.0
beautifulsoup4==4.12.3
biopython==1.85
blinker==1.9.0
cachetools==5.5.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
executing==2.1.0
fasteners==0.19
fonttools==4.55.0
gitdb==4.0.12
GitPython==3.1.44
GridDataFormats==1.0.2
idna==3.10
ipython==8.30.0
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
mda-xdrlib==0.2.0
mdahole2==0.5.0
MDAnalysis==2.8.0
mdurl==0.1.2
mmtf-python==1.1.3
mrcfile==1.5.3
msgpack==1.1.0
narwhals==1.28.0
numpy==2.1.3
networkx==3.4.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathsimanalysis==1.2.0
pillow==11.0.0
prompt_toolkit==3.0.48
protobuf==5.29.3
pure_eval==0.2.3
plotly==5.0.0
pyarrow==19.0.1
pydeck==0.9.1
pygame==2.6.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
rcsbsearchapi==2.0.0
rdkit==2023.9.5
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
scipy==1.14.1
setuptools==75.6.0
six==1.16.0
smmap==5.0.2
soupsieve==2.6
stack-data==0.6.3
streamlit==1.42.2
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
waterdynamics==1.2.0
wcwidth==0.2.13
wheel==0.45.1
widgetsnbextension==4.0.13
